+++ cypress/e2e/happypath/sendfunds_relay.cy.js
@@ -11,8 +11,8 @@ import SafeApiKit from '@safe-global/api-kit'
 import { createSigners } from '../../support/api/utils_ether'
 import { createSafes } from '../../support/api/utils_protocolkit'
 import { contracts, abi_qtrust, abi_nft_pc2 } from '../../support/api/contracts'
-import { getSafes, CATEGORIES } from '../../support/safes/safesHandler.js'
 import * as wallet from '../../support/utils/wallet.js'
+import * as fundSafes from '../../fixtures/safes/funds.json'
 
 const transferAmount = '1'
 
@@ -55,7 +55,7 @@ describe('Send funds with relay happy path tests', { defaultCommandTimeout: 3000
         ls.cookies.acceptedTokenListOnboarding,
       )
     })
-    safesData = await getSafes(CATEGORIES.funds)
+    safesData = fundSafes
     apiKit = new SafeApiKit({
       chainId: BigInt(1),
       txServiceUrl: constants.stagingTxServiceUrl,
